Styled events in weekly based on event time
Events in weekly view are now positioned relative to the the Day/Time table.
Their top position is offset from the nearest hour to their actual start
time and their length accurately reflects the length of the event.
There is no handling of overlapping events and events don't correctly
run over to the next day.