Correct errors in current tests
I discovered during a trip that Cypress was not working with my laptop
dev environment (linux subshell for windows). These commits fix
current tests, but a lot of work remains. None of the event system is
tested, and utility functions are still in desparate need of a refactor,
reorganization, and their own tests.