Event form checks for an id before creation.
Event id weren't being properly passed to the form, causing events to
be duplicated rather than updated.

Now, on submission, the form checks to see if an event has an id to determine
whether a new event is being created or an existing event is being
updated. If an id is found it calls the update event handler otherwise
it calls the event creation handler.