Move Header out of Calendar and change api to Aug
events.json should be in the App folder so the App.js import has been
adjusted. api.js has been bumped to generate events for the month of
august (convienent for development).

Header has been moved outside of the Calendar component- it should not
have been there in the first place as the Calendar has its own header.

Moving the header was the first step in reworking the app routing.
Embedding the router throughout the app will clean up the code a bit
but also break our use of html table. Given that using <table> has
caused nothing but frustration, the html will be rewritten and have
gridcell roles applied instead. Until aria testing is properly performed
it is still a shot in the dark but I figure its better than nothing.