Replace html table with html5 and nest router
Nesting the router in Calendar breaks html table semantics as it returns
a Router (div). Previously we parsed props to manage the change but it
added ugly code and didn't take advantage of our routing library.

Html tables had also already frustrated styling and appeared to offer
very little for built in aria-functionality so I elected to remove it
completely. The css has yet to be rewritten as of this commit so see the
following commit to access a useable site.