Rename date to day and minor Weekly refactor
Calendar and Monthly had already been updated to use the word 'date' to
refer to the numerical day of the month- now Weekly does as well.

Day now refers to the { year, month, date } object used throughout the app.

A const declaration was moved in Weekly/index to directly return a JSX
object from a map which eliminated the explicit return statement.