Encapsulate Header logic into separate components
Section Header generates the separate aria and visual headers. Weekly
view has not been updated because it is about to be completely
re-written. The Calendar's header info has been completely abstracted to
simplify the Calendar component's concerns.

Next will be abstracting the form updating to a context to simplify the
click handlers and to remove components having to generate a partial
event to pass to callbacks on new event creation.

Weekly and Monthly views also need a rewrite in preparation for adding
multi-day event support (and possibly concurrent event display)