Create separate route for event management
Event crud operations now have their own segment of the url. Almost
every file in the application had to be touched to adjust click handlers
and remove the form handling that the Calendar component used to do.

Styling needs work and the Event code is a bit of a mess. If an event is
currently rendered, clicking a different event (or a region with a route
to create an event) does not update the rendered EventForm because the
EventForm is managing its own state. EventForm needs to be rewritten.