Events are listed in each hour under weekly
The redundancy is in preparation for enabling multi-day events. This
also allows long events to show up in any grid-cell in which they occur
rather than only the starting cell (should make aria support easier).

Next will be a style update that should set the height and offset of
each cell in such as manner as to appear continuous across grid-cells.

After the styling is in, the redundant visible names will need to be
hidden from view (excepting events that roll over across days).