Add hacky event deletion.
I think the router redirects are asyncronous causing issues with how
events are being deleted. I believe the next step is to add a component
that wraps event form and manages its render.

For now though, click handlers for event creation really need to be
fixed it makes quick manual tests a pain. Updating the test suite is
also very high priorty, as is a few rounds of refactoring. Regrettably
the hacky solution is likely to stay around for awhile.