Add login link
Have to use an actual <a> tag rather than the routing library because it
reaches out through the proxy server. A full repaint would occur with
the user logging in so not much of a problem for now.

Will update the link text to be dynamic based on status of UserContext
once the context has been implemented.