Add unit tests for SectionHeader
This is the first time using any form of setup function for testing.
The format looks a little different, and I am still unsure if it
would be better to use Jests built in setup functions. All that to say
that other tests may not be updated to this format.