Switch to react-hooks-form library
Form complexity was growing rapidly and error handling was making it
worse. The form is now actually unmanaged, but the code is simplified
and we can now focus on doing what we need to do, which is validating
the form inputs. Specifically, we need to ensure that the start date is
actually before the end date.

Unfortunately, the library seems incompatible with how I have been doing
render testing so I will need to research into how to test the form
properly.