Error messages added with glitchy styling
The error box will grow in width because of the way the messages are
forced to display in a new flex row. Need to consider different position
options for error messages. Before spending too much time restyling, I
also need to look for article on associating error messages with form
inputs for aria. Any required semantic changes could decide on the
proper solution to error styling.