Event form is now modal
The event form styling was inconsistent across browsers and did not
resize well. While fixing these issues, I decided to make it a screen
centered modal. I do not actually like modals, but for now this seemed
like the cleanest solution. Eventually it should appear in a minimizable
box in the bottom right so that the calendar is accessible and multiple
events can be open for editing at a time.