Event creation works through connection loss
Did not update other crud operations for two reasons. The first is that
a queue needs to be set up to execute all offline changes to the account
upon re-connect. Secondly, after learning more about useReducer, I think
the entire context may be re-written with reducer hooks in mind. The
code quality is quite poor due to prototyping the offline caching and it
seems like a much cleaner way to manage the requirements of the event
context.