Abstract event state logic to apiFactory
EventContext has been the cursed file for adding all the various cache
udpating and syncronizing logic. It has become covoluted with react
hook, database, and dexie calls. In order to clean it up, all of the
logic for updating the react hook state has been abstracted into an
apiFactory and a useApi hook. See https://tinyurl.com/uphogfv for the
origin of the abstraction.